fonts: Use fallback variant of Roboto
Harmony is missing some glyphs that users have come to expect on Android,
so use Roboto as a fallback to fix excessive tofu in some text.

Because Harmony's name has been replaced with "Roboto" for compatibility,
the fallback variants are named "Roboto Fallback".

Example of text with missing glyphs in Harmony: ⅙⅐⅛⅑⅒¹½⅓¼⅕

Change-Id: I4affbde067bc51366a84a93b48a9c373567dca05
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>